Renamed User.js to Profile.js, updated protected routes, and created Unauthorized page with return button
