+++ client/src/components/RequireAuth.js
@@ -1,15 +1,16 @@
 import { useLocation, Navigate, Outlet } from "react-router-dom";
 import useAuth from "../hooks/useAuth.js";
 
-const RequireAuth = () => {
+const RequireAuth = ({ allowedRoles }) => {
     const { auth } = useAuth();
     const location = useLocation();
 
     return (
-        auth?.user ? 
-            <Outlet /> 
-        : 
-            <Navigate to="/login" state={{from: location }} replace />
+        auth?.roles?.find(role => allowedRoles?.includes(role))
+            ? <Outlet /> 
+            : auth?.user
+                ?   <Navigate to="/unauthorized" state={{from: location }} replace />
+                :   <Navigate to="/login" state={{from: location }} replace />
     );
 };
